+++ src/NEFDashboard/components/incomeCard/index.jsx
@@ -1,15 +1,19 @@
 import React from "react";
 import { Card, Flex, Row } from "antd";
 import ArrowUp from "../../../assets/newDashboardIcons/arrow-up-green.svg";
+import ArrowDown from "../../../assets/newDashboardIcons/arrow-down-red.svg";
 import InfoMark from "../../../assets/newDashboardIcons/info-circle.svg";
 import { formatTokenBalance } from "../../../utils";
 import { useTranslation } from "react-i18next";
 import { useAuth } from "../../../context/auth/authContext";
+import { isFloat } from "../../../utils";
 import "./incomeCard.css";
 
 const IncomeCard = ({ card, key, isLast = false }) => {
   const { t } = useTranslation();
   const { currencyRate } = useAuth();
+  const isProfit = isFloat(card.percentage) ? card?.percentage > 0 : true;
+
   return (
     <Card
       title={
@@ -31,14 +35,23 @@ const IncomeCard = ({ card, key, isLast = false }) => {
             : card?.number}
         </div>
         <Flex align={"center"} gap={8} className="income-profit-container">
-          <Row align={"middle"} className="income-profit">
+          <Row
+            align={"middle"}
+            className={isProfit ? "income-profit" : "income-loss"}
+          >
             <div className="default-text">
-              +
+              {isProfit ? "+" : ""}
               {card?.percentage == null
                 ? formatTokenBalance(0, 2)
                 : formatTokenBalance(card?.percentage)}
+              %
             </div>
-            <img src={ArrowUp} alt="arrow" width={14} height={14} />
+            <img
+              src={isProfit ? ArrowUp : ArrowDown}
+              alt="arrow"
+              width={14}
+              height={14}
+            />
           </Row>
           <div className="default-text-gray">{t(card?.subText)}</div>
         </Flex>